backend/requirements.txt
# Django Core
Django==5.2.3
djangorestframework==3.16.0

# Database
psycopg2-binary==2.9.10

# Development and Testing
django-filter==25.1
django-extensions==3.2.3

# Environment Management
python-decouple==3.8
django-environ==0.10.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Optional: API Documentation
drf-spectacular==0.27.0

# Optional: Authentication
djangorestframework-simplejwt==5.3.0

# Optional: CORS for frontend integration
django-cors-headers==4.7.0
